Restructure data format gathered from scrapper
